restructure redux data persistance implemenation
